Notifications for ADS state changes

+++ src/test/scala/com/vroste/adsclient/MyManyData.scala
@@ -1,3 +1,3 @@
 package com.vroste.adsclient
 
-case class MyManyData(var1: Int, var2: String, var4: Int, var5: MyStruct)
+case class MyManyData(var1: Short, var2: String, var4: Short, var5: MyStruct)
